requirements.txt
requests~=2.13.0
plaid-python~=2.0.3
